refactor(tests): extract log matchers into dedicated logmatchers package
Move the log matcher implementation from vmhelpers/log_matchers.go and
common_matchers.go into tests/logmatchers/ so they can be imported
without pulling in the full vmhelpers dependency. The vmhelpers package
retains thin wrappers for backward compatibility.
