+++ tests/tls_challenge_test.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/namespaces"
+	"github.com/stackrox/rox/tests/logmatchers"
 	"github.com/stretchr/testify/suite"
 	appsV1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
@@ -96,10 +97,10 @@ func (ts *TLSChallengeSuite) TestTLSChallenge() {
 	ts.logf("Sensor will now attempt connecting via the nginx proxy.")
 
 	ts.waitUntilLog(ts.ctx, s, map[string]string{"app": "sensor"}, sensorContainer, "contain info about successful connection",
-		containsLineMatching(regexp.MustCompile("Info: Add central CA cert with CommonName: 'Custom Root'")),
-		containsLineMatching(regexp.MustCompile("Info: Connecting to Central server "+proxyEndpoint)),
-		containsLineMatching(regexp.MustCompile("Info: Established connection to Central.")),
-		containsLineMatching(regexp.MustCompile("Info: Communication with central started.")),
+		logmatchers.ContainsLineMatching(regexp.MustCompile("Info: Add central CA cert with CommonName: 'Custom Root'")),
+		logmatchers.ContainsLineMatching(regexp.MustCompile("Info: Connecting to Central server "+proxyEndpoint)),
+		logmatchers.ContainsLineMatching(regexp.MustCompile("Info: Established connection to Central.")),
+		logmatchers.ContainsLineMatching(regexp.MustCompile("Info: Communication with central started.")),
 	)
 	waitUntilCentralSensorConnectionIs(ts.T(), ts.ctx, storage.ClusterHealthStatus_HEALTHY)
 }
